added pretrained vww model and refactored repository
